exercise 20.2 change bigdecimal to double

+++ kodilla-patterns_2/src/main/java/com/kodilla/patterns_2/decorator/pizza/Pizza.java
@@ -1,8 +1,6 @@
 package com.kodilla.patterns_2.decorator.pizza;
 
-import java.math.BigDecimal;
-
 public interface Pizza {
-    BigDecimal getCost();
+    double getCost();
     String getDescription();
 }
